Refactor build script
Moved the logic into a function so the module can be imported and used
elsewhere.

Factored out the configurable constants to the top of the file. I also
reworded the descriptions for these to use a more extractable style. We
can probably remove a bunch of the comments in the code since they
explain what/how and not why, but for now they're fine to leave in.

It now calculates the best number of columns to make the resulting grid
a square.

Added a final pass if the user has optipng installed to reduce the
amount of space the image takes up.

Ran black and pylint over it to fix up any issues and use a consistent
style.

Added a shebang and gave it execute privileges so it can just be run
direct.

Added missing newline to the end of file to make it POSIX compliant.

Fixed a file descriptor leak by using a with context block when opening
the image files.

Sort the geets by name to stabilize the order.