Sync files with slack
Fix up a few discrepancies between what gootsmoji are in Slack and which
are in the repo.

Slack is currently missing the following which are in the repo:
- :dead-code-goots:
- :goots-namaste:
- :uno-reverso-blue:
- :uno-reverso-green:
- :uno-reverso-pink:
- :uno-reverso-purple:
- :uno-reverso-yellow:

I moved the gootsmoji we no longer have in slack to an archive folder.
We can delete those if desired. I just wasn't sure if we wanted to keep
or yeet gootsmoji we aren't using anymore.

I also moved the layers out of the goots directory to help tidy the
purpose of each a bit.