Fix toast location from absolute to fixed
I messed up the toast bar location. If it's absolute, that puts it at
the bottom of the page. If it's fixed, it puts it at the bottom of the
viewport. They look the same until you have enough geets to require
scrolling.

I also switched to a flexbox to make the display a little more
interpretable than using an auto margin position.

Manually updated picker list.