Streamline and strengthen build commit commands
Instead of configuring git when creating the commit, we can just pass
the author to the git-commit(1) command. Since we're already using the
-a (--all) flag, it also doesn't make sense to first add files to the
staging are. Additionally, this stage was having problems when the build
doesn't change anything. Now it first tests to see if git has any
changes from the result of the build command to commit.